git-svn-id: svn://123.57.173.36/meijia-parent@1937 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-app/src/main/java/com/simi/action/app/card/CardController.java
@@ -425,11 +425,13 @@ public class CardController extends BaseController {
 			UserMsgSearchVo searchVo = new UserMsgSearchVo();
 			searchVo.setAction("card");
 			searchVo.setParams(cardId.toString());
-			searchVo.setStartTime(TimeStampUtil.getNowSecond());
+//			searchVo.setStartTime(TimeStampUtil.getNowSecond());
 			
 			List<UserMsg> list = userMsgService.selectBySearchVo(searchVo);
 			for (UserMsg item: list) {
-				userMsgService.deleteByPrimaryKey(item.getMsgId());
+				if (item.getServiceTime() > TimeStampUtil.getNowSecond()) {
+					userMsgService.deleteByPrimaryKey(item.getMsgId());
+				}
 			}
 		}
